+++ mod-source-record-manager-server/src/main/java/org/folio/services/JobExecutionService.java
@@ -2,6 +2,8 @@ package org.folio.services;
 
 import io.vertx.core.Future;
 import org.folio.dao.JobExecutionDao;
+import org.folio.dao.JobExecutionFilter;
+import org.folio.dao.util.SortField;
 import org.folio.dataimport.util.OkapiConnectionParams;
 import org.folio.rest.jaxrs.model.InitJobExecutionsRqDto;
 import org.folio.rest.jaxrs.model.InitJobExecutionsRsDto;
@@ -10,6 +12,7 @@ import org.folio.rest.jaxrs.model.JobExecutionDtoCollection;
 import org.folio.rest.jaxrs.model.JobProfileInfo;
 import org.folio.rest.jaxrs.model.StatusDto;
 
+import java.util.List;
 import java.util.Optional;
 
 /**
@@ -22,14 +25,15 @@ import java.util.Optional;
 public interface JobExecutionService {
 
   /**
-   * Returns JobExecutionCollectionDto by the input query
+   * Returns JobExecutionCollectionDto by the input filter
    *
-   * @param query  query string to filter entities
+   * @param filter filter containing conditions by which jobExecutions should be filtered
+   * @param sortFields fields to sort jobExecutions
    * @param offset starting index in a list of results
    * @param limit  maximum number of results to return
    * @return future with JobExecutionCollectionDto
    */
-  Future<JobExecutionDtoCollection> getJobExecutionsWithoutParentMultiple(String query, int offset, int limit, String tenantId);
+  Future<JobExecutionDtoCollection> getJobExecutionsWithoutParentMultiple(JobExecutionFilter filter, List<SortField> sortFields, int offset, int limit, String tenantId);
 
   /**
    * Performs creation of JobExecution and Snapshot entities
@@ -75,12 +79,11 @@ public interface JobExecutionService {
    * to limit the collection param limit should be explicitly specified
    *
    * @param parentId JobExecution parent id
-   * @param query    query string to filter entities
    * @param offset   starting index in a list of results
    * @param limit    maximum number of results to return
    * @return future with collection of child JobExecutions
    */
-  Future<JobExecutionDtoCollection> getJobExecutionCollectionByParentId(String parentId, String query, int offset, int limit, String tenantId);
+  Future<JobExecutionDtoCollection> getJobExecutionCollectionByParentId(String parentId, int offset, int limit, String tenantId);
 
   /**
    * Updates status for JobExecution and calls source-record-storage to update Snapshot status