+++ mod-source-record-manager-server/src/main/java/org/folio/services/JobExecutionsCache.java
@@ -8,11 +8,14 @@ import io.vertx.core.Vertx;
 import org.apache.commons.lang3.tuple.Pair;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.folio.dao.JobExecutionFilter;
+import org.folio.dao.util.SortField;
 import org.folio.rest.jaxrs.model.JobExecutionDtoCollection;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
+import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.TimeUnit;
@@ -27,7 +30,7 @@ public class JobExecutionsCache {
 
   private Integer expireInSeconds;
   private JobExecutionService jobExecutionService;
-  // Pair of tenant id and cql query
+  // Pair of tenant id and conditions from JobExecutionFilter
   private AsyncLoadingCache<Pair<String, String>, Optional<JobExecutionDtoCollection>> cache;
   private Vertx vertx;
 
@@ -40,15 +43,15 @@ public class JobExecutionsCache {
     cache = buildCache();
   }
 
-  public Future<JobExecutionDtoCollection> get(String tenantId, String cqlQuery, int offset, int limit) {
+  public Future<JobExecutionDtoCollection> get(String tenantId, JobExecutionFilter filter, List<SortField> sortFields, int offset, int limit) {
     Promise<JobExecutionDtoCollection> promise = Promise.promise();
-    String uniqueQuery = appendLimitAndOffset(cqlQuery, offset, limit);
+    String uniqueQuery = appendSortAndPagingParams(filter, sortFields, offset, limit);
     cache.get(Pair.of(tenantId, uniqueQuery)).whenComplete((jobExecutionOptional, e) -> {
       if (e == null) {
         if (jobExecutionOptional.isPresent()) {
           promise.complete(jobExecutionOptional.get());
         } else {
-          jobExecutionService.getJobExecutionsWithoutParentMultiple(cqlQuery, offset, limit, tenantId)
+          jobExecutionService.getJobExecutionsWithoutParentMultiple(filter, sortFields, offset, limit, tenantId)
             .onSuccess(ar -> {
               put(tenantId, uniqueQuery, ar);
               promise.complete(ar);
@@ -65,8 +68,9 @@ public class JobExecutionsCache {
     return promise.future();
   }
 
-  private String appendLimitAndOffset(String cqlQuery, int offset, int limit) {
-    return String.format("%s LIMIT %d OFFSET %d", cqlQuery, limit, offset);
+
+  private String appendSortAndPagingParams(JobExecutionFilter filter, List<SortField> sortFields, int offset, int limit) {
+    return String.format("%s sortBy=%s LIMIT %d OFFSET %d", filter.buildCriteria(), sortFields, limit, offset);
   }
 
   private void put(String tenantId, String cqlQuery, JobExecutionDtoCollection jobExecutionCollection) {